requirements.txt
pycryptodome==3.18.0
PyQt5==5.15.9 